Fix import error from "asym.py"
When importing the last asymmetry package, the __init__.py script simply imports everything from asym.py.  But in the current version, the import command includes the ".py" at the end of the filename, which flags an ImportError when using the packge.

The fix above on line 1 of the __init__.py script simply removes the ".py" at the end of the import statement to make importing the package as simple as the example claims from the README file / github home.