Added least_asymmetry package to site-packages
currently, to use least_asymmetry in the way that the README and github home page describe, the user must have the least_asymmetry github repo in their `PYTHONPATH`.  

This PR creates a sub-routine in the `setup.py` file which adds the `least_asymmetry` package to the standard `site-packages` sub-directory taken from the `PYTHONPATH` variable.

The edits made here 
1. locate the python site-packages directory
2. identify if the subdirectory `/path/to/site-packages/least_asymmetry` exists
3a. if the sub-directory does not exist, then the setup.py script creates it
3b. if the sub-directory does exist, then the setup.py script requests to overwrite the files within
4. the setup.py script then copies the __init__.py and asym.py (the only import from __init__.py) from the github repo to the newly created site-packages directory.

This fixes needing to add the github repo to the `PYTHONPATH` environment variables; this assumes that the site-packages directory is itself in the `PATH` env variable -- a very standard procedure.